math.c implemented

+++ Assignment/math.h
@@ -18,6 +18,6 @@
 #define MATH_H_
 
 //Part of your documentation should be listing the valid inputs and outputs for the functions you create.
-int math(int num1, int num2, char Operator);
+float math(int num1, int num2, char Operator);
 
 #endif /* MATH_H_ */
